requirements.txt
clip @ git+https://github.com/openai/CLIP.git
pymilvus
tqdm
streamlit
pillow